+++ codegen/tests/trait_and_type_generic_param.rs
@@ -3,7 +3,7 @@ use delegation::delegate;
 #[delegate(for(
     for<U> Case2<U>
     where
-        U: Named<N> + 'static;
+        U: Named<N> + 'static,
 ))]
 trait Named<N> {
     fn name(&self) -> N;
@@ -19,7 +19,7 @@ impl Named<String> for User {
 #[delegate(derive(
     for<N> Named<N>
     where
-        U: Named<N> + 'static;
+        U: Named<N> + 'static,
 ))]
 enum Case1<U> {
     User(U),
@@ -31,7 +31,7 @@ struct Case2<U>(U);
 #[delegate(derive(
    Named<String>
    where
-       U: Named<String> + 'static;
+       U: Named<String> + 'static,
 ))]
 enum Case3<U> {
     Case1(Case1<U>),